backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.20
pg8000==1.30.3
alembic==1.12.0
groq==0.4.2
redis==5.0.0
python-dotenv==1.0.0
google-generativeai==0.8.3
